Fix a bug about flush bits when we will have an integer overflow
Currently, decompress try to avoid an integer overflow while it encodes the
given queue. In this case, it calls flush_bits which wants emit `hold` as much
as possible. When it safe to store bits into hold, flush_bits should come back
to the hot-loop to continue to consume the queue and encode the flow.

However, before this patch, flush_bits return [`Ok] which means end of the
process to the user. Instead, we put a k_continue function which recall the
hot-loop as expected.

We removed a failwith (reached an other time) and just try to call and recall
flush_bits while we can not store given bits to the encoder.

Again, this specific process slow-down the encoding but it does that into a
safe-way. At least, the case where flush_bits is called and when it should not
return [`Ok] is when we compress paper1 with GZip layer.