Really call the internal continuation of Zlib.Inf
It's not logic to have a k into the state and never call it then. Currently,
according interface and implementation, we should have some bug - mostly when we
have small outputs - because some processes trust on the ability to record a
specific continuation when, by the fact, it will be never called.

This patch is minor but it wants to fix it.